+++ NEWS.md
@@ -4,7 +4,9 @@
 * Made bracket corners join neatly.
 * Supported As Is normalised coordinates (e.g. `xintercept = I(0.5)`).
 * Supported vectorised style arguments.
-* Removed minor argument from `axis_ticks`.
+* Removed `minor` argument from `axis_ticks`.
+* Added `layout` argument to support panel annotations on facet plots.
+* Added `sec_axis_spacer` function to provide space.
 * Removed `theme_sec_axis_text`.
 
 # ggscribe 0.1.1